spawn function in engine
