+++ pgdrive/component/map/base_map.py
@@ -4,7 +4,7 @@ import logging
 import numpy as np
 
 from pgdrive.component.algorithm.BIG import BigGenerateMethod
-from pgdrive.component.base_class.base_runable import BaseRunnable
+from pgdrive.base_class.base_runnable import BaseRunnable
 from pgdrive.component.blocks.first_block import FirstPGBlock
 from pgdrive.component.blocks.pg_block import PGBlock
 from pgdrive.component.road.road import Road
