+++ pgdrive/component/static_object/base_static_object.py
@@ -2,7 +2,7 @@ from typing import Sequence, Tuple
 
 import numpy as np
 
-from pgdrive.component.base_class.base_object import BaseObject
+from pgdrive.base_class.base_object import BaseObject
 
 LaneIndex = Tuple[str, str, int]
