+++ pgdrive/envs/marl_envs/marl_bottleneck.py
@@ -78,7 +78,7 @@ class MultiAgentBottleneckEnv(MultiAgentPGDrive):
 
         if self.current_map is None:
             self.seed(map_config["seed"])
-            new_map = self.engine.map_manager.spawn_object(
+            new_map = self.engine.spawn_object(
                 MABottleneckMap, map_config=map_config, random_seed=self.current_seed
             )
             self.engine.map_manager.load_map(new_map)
