+++ pgdrive/manager/base_manager.py
@@ -1,6 +1,4 @@
-from typing import Callable, Optional
-
-from pgdrive.component.base_class.randomizable import Randomizable
+from pgdrive.base_class.randomizable import Randomizable
 
 
 class BaseManager(Randomizable):
@@ -12,46 +10,6 @@ class BaseManager(Randomizable):
         assert engine_initialized(), "You should not create manager before the initialization of BaseEngine"
         self.engine = get_engine()
         Randomizable.__init__(self, self.engine.global_random_seed)
-        self._spawned_objects = dict()
-
-    def spawn_object(self, object_class, *args, **kwargs):
-        """
-        Call this func to spawn one object
-        :param object_class: object class
-        :param kwargs: class init parameters
-        :return: object spawned
-        """
-        if "random_seed" not in kwargs:
-            kwargs["random_seed"] = self.randint()
-        obj = object_class(*args, **kwargs)
-        self._spawned_objects[obj.id] = obj
-        return obj
-
-    def get_objects(self, filter_func: Optional[Callable] = None):
-        """
-        Return objects spawned and managed by this manager, default all objects
-        Since we don't expect a iterator, and the number of objects is not so large, we don't use built-in filter()
-        :param filter_func: a filter function, only return objects satisfying this condition
-        :return: return all objects or objects satisfying the filter_func
-        """
-        res = dict()
-        for id, obj in self._spawned_objects.items():
-            if filter_func is None or filter_func(obj):
-                res[id] = obj
-        return res
-
-    def clear_objects(self, filter_func: Optional[Callable] = None):
-        """
-        Destroy all self-generated objects or objects satisfying the filter condition
-        Since we don't expect a iterator, and the number of objects is not so large, we don't use built-in filter()
-        """
-        exclude = []
-        for id, obj in self._spawned_objects.items():
-            if filter_func is None or filter_func(obj):
-                obj.destroy()
-            exclude.append(id)
-        for id in exclude:
-            self._spawned_objects.pop(id)
 
     def before_step(self, *args, **kwargs) -> dict:
         """
@@ -94,6 +52,4 @@ class BaseManager(Randomizable):
         """
         Destroy manager
         """
-        self.clear_objects()
-        self._spawned_objects = None
         self.engine = None
