+++ pgdrive/manager/policy_manager.py
@@ -12,15 +12,13 @@ class PolicyManager(BaseManager):
         Some policy might be stateful, for example a LSTM-based neural policy network. We need to reset the states
         of all policies here.
         """
-        for p in self._spawned_objects.values():
-            p.reset()
-        print('222')
+        return
 
     def register_new_policy(self, policy_class, vehicle, traffic_manager, *args, **kwargs):
         # e = get_engine()
         # TODO: We should have a general "vehicle manager". Then we can get the BaseVehicle instance from
         #  engine according to agent_name! Here is only a workaround.
-        policy = self.spawn_object(policy_class, vehicle=vehicle, traffic_manager=traffic_manager, *args, **kwargs)
+        policy = self.engine.spawn_object(policy_class, vehicle=vehicle, traffic_manager=traffic_manager, *args, **kwargs)
         policy_name = policy.name
         self._policy_name_to_vehicle[policy_name] = vehicle
         self._vehicle_name_to_policy[vehicle.name] = policy
@@ -32,9 +30,3 @@ class PolicyManager(BaseManager):
         if vehicle_name not in self._vehicle_name_to_policy:
             return None
         return self._vehicle_name_to_policy[vehicle_name]
-
-    def destroy(self):
-        for p in self._spawned_objects.values():
-            if hasattr(p, "destroy"):
-                p.destroy()
-        super(PolicyManager, self).destroy()
