+++ pgdrive/tests/vis_env/vis_pgdrive_env.py
@@ -58,18 +58,18 @@ if __name__ == "__main__":
         info["fuel"] = env.vehicle.energy_consumption
         # env.render(text={"heading_diff": env.vehicle.heading_diff(env.vehicle.lane)})
         assert env.observation_space.contains(o)
-        env.render(
-            text={
-                "reward": r,
-                "lane_index": env.vehicle.lane_index,
-                "dist_to_left": env.vehicle.dist_to_left_side,
-                "dist_to_right": env.vehicle.dist_to_right_side,
-                "out_of_route": env.vehicle.out_of_route,
-                "current_seed": env.current_seed,
-                "car_heading": env.vehicle.heading_theta,
-                "lane_heading": env.vehicle.lane.heading_at(0)
-            }
-        )
+        # env.render(
+        #     text={
+        #         "reward": r,
+        #         "lane_index": env.vehicle.lane_index,
+        #         "dist_to_left": env.vehicle.dist_to_left_side,
+        #         "dist_to_right": env.vehicle.dist_to_right_side,
+        #         "out_of_route": env.vehicle.out_of_route,
+        #         "current_seed": env.current_seed,
+        #         "car_heading": env.vehicle.heading_theta,
+        #         "lane_heading": env.vehicle.lane.heading_at(0)
+        #     }
+        # )
         if d:
             print("Reset")
             env.reset()
